rename camelCase files
