+++ package.json
@@ -38,7 +38,6 @@
     "gatsby-transformer-remark": "2.6.22",
     "gatsby-transformer-sharp": "2.2.14",
     "gatsby-transformer-yaml": "2.2.9",
-    "lodash": "^4.17.15",
     "prismjs": "1.17.1",
     "prop-types": "15.7.2",
     "react": "16.9.0",
