+++ src/components/shared/form/index.tsx
@@ -1,4 +1,4 @@
 export { Form } from './forms';
-export { CreateProjectForm } from './createProject';
+export { CreateProjectForm } from './create-project';
 export { ErrorMessage } from './error-msg';
 export * from './controls';
